update creation slot action
